pyproject.toml
[project]
name = "zotero-arxiv-daily"
version = "0.3.5"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.1.3",
    "llama-cpp-python>=0.3.2",
    "loguru>=0.7.2",
    "pyzotero>=1.5.25",
    "scikit-learn>=1.5.2",
    "sentence-transformers>=3.3.1",
    "openai>=1.57.0",
    "gitignore-parser>=0.1.11",
    "tiktoken>=0.8.0",
    "python-dotenv>=1.0.1",
    "feedparser>=6.0.11",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
]
